Move note parser to sep folder
