+++ src/main/lib/parser/convert-scan.ts
