+++ src/main/lib/parser/sqlite/index.js
