+++ src/main/lib/parser/sqlite/utils.js
